Modify package for a PyPI release
This is a major reshuffling of the package. In order to distribute the
calibration sample paths with the package the paths were placed in a
JSON instead of a number of separate files and symlinks. Various aspects
related to testing had to be changed as well to allow the user to run
tests on the installed package regardless of the current directory.
main() was added to each package that is to be executable as a script.
This pairs with a setup.cfg console_scripts to automatically add the
executables in the correct path.