Disable line that replicates old PIDCalib behavior
The original PIDCalib assigned bins with no events in the total
histogram an efficiency of zero. This should not come up often and the
user should be warned about it. In any case it does not seem right -
we assign the bin a NaN. This might cause slightly different results
when using a sample and binning that lead to such empty bins.